requirement.txt
streamlit
openai
python-dotenv
pypdf
fpdf
